requirements.txt
aiohttp==3.7.4.post0
anyio==3.3.4
asgiref==3.4.1
astroid==2.8.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
bidict==0.21.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
ecdsa==0.17.0
fastapi==0.70.0
fastapi-socketio==0.0.8
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.20.0
idna==3.3
isort==5.9.3
Jinja2==3.0.2
lazy-object-proxy==1.6.0
limits==1.5.1
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.2.0
orjson==3.6.4
passlib==1.7.4
platformdirs==2.4.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pylint==2.11.1
python-engineio==4.2.1
python-jose==3.3.0
python-multipart==0.0.5
python-socketio==5.4.1
pytz==2021.3
redis==3.5.3
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
slowapi==0.1.5
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
toml==0.10.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
uvloop==0.16.0
wrapt==1.13.2
yarl==1.7.0
